refactor imports & utils
