+++ predicate/predicate_bytes.go
@@ -44,30 +44,6 @@ func UnpackPredicate(paddedPredicate []byte) ([]byte, error) {
 	return trimmedPredicateBytes[:len(trimmedPredicateBytes)-1], nil
 }
 
-// HashSliceToBytes serializes a []common.Hash into a tightly packed byte array.
-func HashSliceToBytes(hashes []common.Hash) []byte {
-	bytes := make([]byte, common.HashLength*len(hashes))
-	for i, hash := range hashes {
-		copy(bytes[i*common.HashLength:], hash[:])
-	}
-	return bytes
-}
-
-// BytesToHashSlice packs [b] into a slice of hash values with zero padding
-// to the right if the length of b is not a multiple of 32.
-func BytesToHashSlice(b []byte) []common.Hash {
-	var (
-		numHashes = (len(b) + 31) / 32
-		hashes    = make([]common.Hash, numHashes)
-	)
-
-	for i := range hashes {
-		start := i * common.HashLength
-		copy(hashes[i][:], b[start:])
-	}
-	return hashes
-}
-
 // GetPredicateResultBytes returns the predicate result bytes from the extra data.
 // If the extra data does not contain predicate result bytes, an error is returned.
 func GetPredicateResultBytes(extraData []byte) ([]byte, error) {
