+++ utils/bytes.go
@@ -3,6 +3,8 @@
 
 package utils
 
+import "github.com/ethereum/go-ethereum/common"
+
 // IncrOne increments bytes value by one
 func IncrOne(bytes []byte) {
 	index := len(bytes) - 1
@@ -15,4 +17,28 @@ func IncrOne(bytes []byte) {
 			index--
 		}
 	}
+}
+
+// HashSliceToBytes serializes a []common.Hash into a tightly packed byte array.
+func HashSliceToBytes(hashes []common.Hash) []byte {
+	bytes := make([]byte, common.HashLength*len(hashes))
+	for i, hash := range hashes {
+		copy(bytes[i*common.HashLength:], hash[:])
+	}
+	return bytes
+}
+
+// BytesToHashSlice packs [b] into a slice of hash values with zero padding
+// to the right if the length of b is not a multiple of 32.
+func BytesToHashSlice(b []byte) []common.Hash {
+	var (
+		numHashes = (len(b) + 31) / 32
+		hashes    = make([]common.Hash, numHashes)
+	)
+
+	for i := range hashes {
+		start := i * common.HashLength
+		copy(hashes[i][:], b[start:])
+	}
+	return hashes
 }
